+++ .gitignore
@@ -23,3 +23,4 @@ dist-ssr
 *.sln
 *.sw?
 coverage
+.next
