--- next.config.mjs
@@ -1,11 +1,0 @@
-import createNextIntlPlugin from 'next-intl/plugin';
-
-const withNextIntl = createNextIntlPlugin();
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  output: 'export',
-  distDir: './dist'
-};
-
-export default withNextIntl(nextConfig);
