+++ next.config.ts
@@ -1,0 +1,17 @@
+import { NextConfig } from 'next';
+import createNextIntlPlugin from 'next-intl/plugin';
+
+const nextConfig: NextConfig = {
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'rickandmortyapi.com',
+        pathname: '/api/character/**',
+      },
+    ],
+  },
+};
+
+const withNextIntl = createNextIntlPlugin();
+export default withNextIntl(nextConfig);
