--- src/app/@details/default.tsx
@@ -1,3 +1,0 @@
-export default function EmptyDetails() {
-  return <p className="text-gray-500">Choose a card</p>;
-}
