--- src/app/@details/page.tsx
@@ -1,3 +1,0 @@
-export default function DetailsPage() {
-  return <></>;
-}
