+++ src/app/[locale]/@details/[id]/default.tsx
@@ -1,0 +1,3 @@
+export default function DefaultId() {
+  return null;
+}
