+++ src/app/[locale]/@details/default.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <p>404 Go Home</p>;
+}
