+++ src/app/[locale]/@details/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return null;
+}
