+++ src/app/[locale]/about/page.tsx
@@ -1,0 +1,5 @@
+import { AboutPage } from '../../../pages/AboutPage';
+
+export default function About() {
+  return <AboutPage />;
+}
