+++ src/app/[locale]/default.tsx
@@ -1,5 +1,5 @@
 import { MainPage } from '../../pages/MainPage';
 
-export default function Page() {
+export default function DefaultId() {
   return <MainPage />;
 }
