--- src/app/page.tsx
@@ -1,5 +1,0 @@
-import { MainPage } from '../pages/MainPage';
-
-export default function Page() {
-  return <MainPage />;
-}
